chore(arch): enforce clean-architecture invariants in CI
Move LlmEvent union to core/events.py so llm→core stays unidirectional,
add SmartRoutingProtocol in core/ to eliminate two TYPE_CHECKING imports
from core→llm, wire import-linter with a layered contract, and run it in
CI alongside the existing gates.

Closes #651
