+++ .github/workflows/ci.yml
@@ -31,6 +31,9 @@ jobs:
       - name: Check file length
         run: bash tools/check_file_length.sh
 
+      - name: Enforce import layers
+        run: uv run lint-imports
+
       - name: Typecheck
         run: uv run pyright
