+++ pyproject.toml
@@ -133,6 +133,7 @@ max-args = 5
 
 [dependency-groups]
 dev = [
+    "import-linter>=2.0",
     "pip-audit>=2.10.0",
     "pip-licenses>=5.5.1",
     "pre-commit>=4.5.1",
