+++ src/lyra/llm/__init__.py
@@ -1,5 +1,6 @@
+from lyra.core.events import LlmEvent, ResultLlmEvent, TextLlmEvent, ToolUseLlmEvent
+
 from .base import LlmProvider, LlmResult
-from .events import LlmEvent, ResultLlmEvent, TextLlmEvent, ToolUseLlmEvent
 from .registry import ProviderRegistry
 
 __all__ = [
