+++ src/lyra/llm/drivers/cli.py
@@ -8,8 +8,8 @@ from pathlib import Path
 
 from lyra.core.agent_config import ModelConfig
 from lyra.core.cli_pool import CliPool
+from lyra.core.events import LlmEvent
 from lyra.llm.base import LlmResult
-from lyra.llm.events import LlmEvent
 
 log = logging.getLogger(__name__)
