+++ src/lyra/llm/smart_routing.py
@@ -13,29 +13,17 @@ import re
 import time
 from collections import deque
 from collections.abc import AsyncIterator
-from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any
 
 from lyra.core.agent_config import Complexity, ModelConfig, SmartRoutingConfig
+from lyra.core.events import LlmEvent
+from lyra.core.smart_routing_protocol import RoutingDecision
 from lyra.llm.base import LlmProvider, LlmResult
-from lyra.llm.events import LlmEvent
 
 if TYPE_CHECKING:
     from lyra.core.message import InboundMessage
 
 log = logging.getLogger(__name__)
-
-
-@dataclass(frozen=True)
-class RoutingDecision:
-    """Record of a single routing decision."""
-
-    complexity: Complexity
-    original_model: str
-    routed_model: str
-    reason: str
-    timestamp: float
-    message_preview: str
 
 
 # ---------------------------------------------------------------------------
