+++ tests/core/test_cli_streaming_lifecycle.py
@@ -7,7 +7,7 @@ import json
 from unittest.mock import AsyncMock, MagicMock
 
 from lyra.core.cli_protocol import StreamingIterator, send_and_read_stream
-from lyra.llm.events import ResultLlmEvent, TextLlmEvent
+from lyra.core.events import ResultLlmEvent, TextLlmEvent
 
 from .conftest import (
     ASSISTANT_INTERMEDIATE_LINE,
